extract shared telemetry to avoid duplication
