+++ workspaces/x2a/plugins/x2a/src/components/PhaseDetails.tsx
@@ -32,8 +32,6 @@ import {
   Phase,
 } from '@red-hat-developer-hub/backstage-plugin-x2a-common';
 
-import millify from 'millify';
-
 import { useTranslation } from '../hooks/useTranslation';
 import { useLogStream } from '../hooks/useLogStream';
 import { useClientService } from '../ClientService';
@@ -46,7 +44,7 @@ import {
   humanizeDate,
   secondsBetween,
 } from './tools';
-import { PhaseTelemetry } from './PhaseTelemetry';
+import { TelemetrySection } from './PhaseTelemetry';
 import { PhaseStatus } from './PhaseStatus';
 
 const useStyles = makeStyles(theme => ({
@@ -205,15 +203,6 @@ export const PhaseDetails = (
 
   const canRunPhase = phase?.status !== 'running';
 
-  const tokenTotals = useMemo(() => {
-    const all = Object.values(phase?.telemetry?.agents ?? {});
-    if (all.length === 0) return undefined;
-    return {
-      inputTokens: all.reduce((s, a) => s + (a.inputTokens ?? 0), 0),
-      outputTokens: all.reduce((s, a) => s + (a.outputTokens ?? 0), 0),
-    };
-  }, [phase?.telemetry]);
-
   const fetchLog = useCallback(
     () =>
       phaseName === 'init'
@@ -352,33 +341,7 @@ export const PhaseDetails = (
         </Grid>
       )}
 
-      {phase?.telemetry && (
-        <>
-          <Grid item xs={12}>
-            <Grid container alignItems="baseline" spacing={2}>
-              <Grid item>
-                <Typography variant="h6">
-                  {t('modulePage.phases.telemetry.title')}
-                </Typography>
-              </Grid>
-              {tokenTotals && (
-                <Grid item>
-                  <Typography variant="body2" color="textSecondary">
-                    {millify(tokenTotals.inputTokens)}{' '}
-                    {t('modulePage.phases.telemetry.totalInputTokens')}
-                    {' · '}
-                    {millify(tokenTotals.outputTokens)}{' '}
-                    {t('modulePage.phases.telemetry.totalOutputTokens')}
-                  </Typography>
-                </Grid>
-              )}
-            </Grid>
-          </Grid>
-          <Grid item xs={12}>
-            <PhaseTelemetry telemetry={phase?.telemetry} />
-          </Grid>
-        </>
-      )}
+      <TelemetrySection telemetry={phase?.telemetry} />
     </Grid>
   );
 };
